+++ source/useCases/lintFs/reporter/print.js
@@ -1,0 +1,26 @@
+const print = ({
+  logBatchArguments,
+  logArguments,
+  logger,
+  shouldReport,
+}) => {
+  if (shouldReport()) {
+    const [
+      , titleError,
+    ] = logger.log(...logArguments);
+    if (titleError) {
+      return titleError;
+    }
+
+    const [
+      , batchError,
+    ] = logger.logBatch(...logBatchArguments);
+    if (batchError) {
+      return batchError;
+    }
+  }
+
+  return '';
+};
+
+export default print;
